Prepare 0.4.8 build 4 with reliable automatic Codex connection
