+++ Configs/Widget.xcconfig
@@ -3,9 +3,9 @@
 PRODUCT_NAME = QuotaViewWidgetExtension
 QUOTAVIEW_WIDGET_BUNDLE_ID = com.quotaview.menubar.widget
 PRODUCT_BUNDLE_IDENTIFIER = $(QUOTAVIEW_WIDGET_BUNDLE_ID)
-MARKETING_VERSION = 0.4.7
-CURRENT_PROJECT_VERSION = 22
-QUOTAVIEW_DISPLAY_BUILD_NUMBER = 3
+MARKETING_VERSION = 0.4.8
+CURRENT_PROJECT_VERSION = 26
+QUOTAVIEW_DISPLAY_BUILD_NUMBER = 4
 
 QUOTAVIEW_APP_GROUP_ID = BUUH229D5Q.com.quotaview.shared
