+++ README.zh-CN.md
@@ -42,13 +42,15 @@
 
 QuotaView 开源、轻量，并以本地处理为核心。新版 Codex 首次启动即可使用，无需配置 Hook；额度与用量则保留在菜单面板和原生小组件中，需要时一次点击即可查看。
 
-## 0.4.7 Build 3：完整的灵动岛文字与自定义代理
+## 0.4.8 Build 4：清晰的首次连接与可靠的任务恢复
 
-Build 3 修复了 **27%** 等数值的百分号消失，以及文字刚好放下时被误截断的问题；同时修正代理协议控件的右对齐，并为设置侧栏加入彩色圆角图标。
+> 发布准备中：0.4.8 正式签名安装包尚未发布，上方下载入口仍为当前稳定版。
 
-**设置 → 代理设置** 支持无需账号密码的 HTTP / SOCKS5 地址，提供连接测试、保存与恢复默认。默认关闭，仅影响 QuotaView 的额度和账户用量查询，灵动岛本地活动与当前任务 Token 继续使用原有数据来源。
+自动连接现在分别显示首次等待、记录可读、已收到活动和读取错误；可在设置中重新检查或选择 Codex 数据目录。兼容 Hook 为可选功能，打开设置只检查状态，未信任的旧安装也可手动移除。
 
-本版替代已撤回的 0.4.7 Build 2，并加入稳定版自动更新。完成后的缩小和隐藏计时保持不变。
+本版修复文件替换后的任务恢复、连接状态订阅遗漏及候选分页时内部任务过滤失效，按职责整理连接、恢复和 Hook 生命周期代码，并统一设置图标内边距。
+
+继续提供 HTTP/SOCKS5 代理设置和长期保留的灵动岛开发台。额度、小组件数据格式以及完成后的缩小和隐藏计时保持原有规则。
 
 ## Codex 灵动岛
