+++ Support/Info.plist
@@ -17,11 +17,11 @@
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleShortVersionString</key>
-    <string>0.4.7</string>
+    <string>0.4.8</string>
     <key>CFBundleVersion</key>
-    <string>22</string>
+    <string>26</string>
     <key>QuotaViewDisplayBuildNumber</key>
-    <string>3</string>
+    <string>4</string>
     <key>LSMinimumSystemVersion</key>
     <string>14.0</string>
     <key>LSApplicationCategoryType</key>
