+++ docs/design/quotaview-codex-first-connection-0.4.8.md
@@ -1,0 +1,131 @@
+# 0.4.8 首次连接引导修复
+
+Spec ID：`QV-FIX-CODEX-FIRST-CONNECTION-019`  
+状态：`Accepted / Verifying`，2026-09-12。用户批准首次连接新方案，版本保持 0.4.8。
+
+基线为 0.4.7 Build 3 / internal 22。用户在全新安装的 Intel Mac 上确认：
+Hook 尚未信任，但运行 Codex 桌面任务后灵动岛可以显示。代码确认自动任务流
+尚在发现时，设置页沿用 Hook 的“未启用 / 连接 Codex”引导。
+
+## 要求与边界
+
+- FIRST-01：首次没有任务记录时显示“等待 Codex 本地任务”，说明运行任务后
+  自动连接；主连接区不能触发 Hook 安装、安全确认或重启。
+- FIRST-02：自动任务流与 Hook 配置分别保存状态。本地自动连接成功不能将
+  未信任的 Hook 显示为已连接；Hook 未信任或配置失败不能覆盖有效自动连接。
+- FIRST-03：Hook 操作位于默认折叠的“兼容选项：Hook”中，由用户显式操作。
+  保留现有安装、信任、重启、事件确认与卸载；停用 Hook 不清空本地任务状态。
+- FIRST-04：断连提示岛仅响应当前会话显式兼容配置操作，且自动任务流或
+  Hook 已连接时不覆盖正常任务；已有未完成 Hook 配置不会在启动时抢占显示。
+- FIRST-05：自动通道禁用时如实显示禁用；发现不到记录只表示等待，不伪造
+  成功或猜测为权限错误。现有轮询持续发现后续生成的任务，无需重启应用。
+
+Build 3 新增已批准要求：
+
+- FIRST-06：连接健康独立表达检查中、无记录、可读就绪、收到活动、读取失败、
+  明确不支持的记录结构与禁用。普通未知事件和静默不能被判为故障。
+- FIRST-07：数据库只用于加速；合并去重有界目录扫描结果，旧索引不阻断新任务。
+- FIRST-08：历史未结束记录仅恢复上下文；新的有效活动或监测启动后带真实时间戳
+  的新任务才能激活显示。完成事件不恢复历史活动，也不凭超时伪造完成。
+- FIRST-09：提供只读重新检查、选择数据目录和恢复默认目录；切换时清除旧目录
+  的活动缓存，任务读取和归属识别使用同一目录。额度和账户配置保持独立。
+- FIRST-10：启动与查看设置只检查 Hook，不启用特性、不重装或改写配置。
+  已有 QuotaView Hook 在未信任、失败、待重启等状态也可手动移除。
+
+保留 Token、内部任务过滤、终态语义、代理、额度、Widget 和渲染器约束。
+不自动修改用户现有 Codex 配置或信任，不增加临时 UI QA 入口。
+当前开发目标为 0.4.8 Build 4 / internal 26；仅本地开发，未获准发布或进入 appcast。
+
+Build 2 隔离审计已确认：目录读取失败被显示为等待、非空旧数据库遗漏新文件、
+旧未结束任务被当作活动。审计证据位于 `dist/verification/build2/first-install-audit/`。
+Build 3 必须回归以上场景及恢复、目录切换、未知格式、Hook 移除和现有 Token 测试。
+
+## Build 1 验证
+
+覆盖首次空目录 → 生成真实格式任务文件 → 自动连接、禁用通道、未信任 Hook
+与自动连接并存、Hook 失败、手动配置、重新断连、停用 Hook 后保留任务。
+运行 Swift 全量测试与 Universal Release 无签名构建，核对版本及资源。
+中英文设置、折叠兼容操作及真实新装机器由产品所有者手动验收。
+
+结果：Swift 201 项、0 失败、2 项既有 opt-in 跳过；4 项新增连接回归通过。
+空 Codex 目录测试通过生产 Store 和本地文件读取器，验证目录生成后自动连接、
+活动展示与真实格式完成事件；该测试未安装 Hook、未启动 Codex 或写账户配置。
+禁用通道及停止后再启动测试也通过。没有代替 Intel 设备上的 0.4.8 实机验收。
+
+最终源码的 Swift 全量复核为 201 项、0 失败、2 项 opt-in 跳过。
+Universal Release 无签名构建通过；App / Widget 均为 0.4.8 / Build 1 / internal 23，
+App、Widget、Core、Hook 均含 x86_64 + arm64；AppIcon.icns 与 Assets.car 存在。
+`git diff --check` 通过，生产源码无临时模拟或 UI QA 注入。
+证据位于 `dist/verification/build1/{swift-tests.log,universal-release.log,artifact-check.json}`。
+未重新执行无改动的真实 120 秒收起周期与已安装 Codex 的代理 opt-in 测试。
+视觉、交互和新装设备验证仍等待用户验收；未签名、公证或发布。
+当前工作与公开基线见 [Handoff](../../HANDOFF.md)。
+
+## Build 2 设置图标边距修正
+
+用户反馈彩色设置图标的内边距不一致。复用既有设置组件规范，采用固定
+20×20pt 底板、14×14pt 符号区域和 3pt 最小边距，符号等比缩放并居中。
+保留原符号、配色、5pt 圆角、原生 Label/List 交互和辅助功能。
+固定符号区域消除不同 SF Symbol 仅按字号渲染产生的宽度差异；长方形符号
+允许上下留白更多，不拉伸形状。该局部修正没有新的业务边界。
+
+Build 2 / internal 24；Swift 201 项、0 失败、2 项既有 opt-in 跳过；
+Universal Release、App/Widget 版本、四个主要二进制的双架构、正式图标和
+Asset Catalog 检查通过，`git diff --check` 通过。
+本轮为局部图标几何调整，没有新增镜像实现的测试或自动截图入口。
+已打开不包含 Widget 的独立 ad-hoc 调试副本，视觉效果等待用户验收。
+证据与路径见 `dist/verification/build2/`；仍未推送或发布。
+
+## Build 3 实现与验收
+
+用户于 2026-09-12 批准新方案。连接健康和任务生命周期独立；本地历史回放
+增加“待新活动确认”的上下文标记，由 Store 缓冲，收到增量活动后恢复 Token
+和计划。只在监测启动后、距发现不超过五秒的有效新任务时间戳出现时直接激活；
+缺失或无效时间戳不走这个入口，历史终态不会唤醒旧任务。
+
+目录遍历每批最多 512 个条目，持续保留遍历游标；本地与 UTC 当天/前一天有
+各 128 条目的快速发现入口，保留最近 1024 个文件并与数据库候选合并。
+文件元数据按 4KB 读取到首行，仍保留原有 1MiB 首行上限。数据库中明确标记
+为内部任务的路径不通过目录补充绕过过滤。旧/迁移/缺失数据库允许目录接管。
+
+只读重查保留现有游标和任务，不调用安装器。用户选择包含可读 sessions 的
+数据目录后，自动读取、共享 Socket、标题与归属识别切换到同一目录，旧状态
+被清空。恢复默认使用 CODEX_HOME 或 ~/.codex；保存选项只写 QuotaView 偏好。
+已有 Hook 不因自动目录变化迁移，其配置目录在兼容详情独立显示。
+Hook 启动检查不再触发特性启用或安装，检查失败仍可移除已知 QuotaView Hook。
+
+新增隔离回归覆盖历史恢复/迟到完成、目录与文件权限恢复、未知与不兼容元数据、
+旧数据库和旧 schema、分页继续扫描、内部任务过滤、目录持久化/重置、Hook
+只读检查及保留其他处理器的移除。移除后的迟到 Hook 事件直接确认并丢弃，
+不会重试污染队列或重新激活任务。
+
+最终 Swift 212 项、0 失败、2 项既有 opt-in 跳过；首次连接 15 项通过。
+Universal Release 无签名构建通过；App/Widget 为 0.4.8 Build 3 / internal 25，
+四个主要二进制均含 arm64 + x86_64；正式图标和 Asset Catalog 检查通过。
+代码/状态/中英文分支及 diff 检查完成，无临时模拟或 UI QA 入口。
+已启动独立 ad-hoc 副本（不注册新的 Widget），stderr 为空；启动前后 Codex
+hooks.json / config.toml 指纹一致，原安装版 Widget 进程保持运行。
+证据：`dist/verification/build3/{swift-tests.log,universal-release.log,artifact-check.json,debug-launch-result.json}`。
+真实首次安装和视觉交互仍等待用户验收；没有提交、推送、发布或修改 appcast。
+
+## Build 4 全链路审计与职责整理
+
+用户要求全面审计相关逻辑，防止修补形成重复状态与失控依赖。本轮复用
+FIRST-01～10，补充实现约束：
+
+- FIRST-11：连接健康具有单一状态来源，移除不再被生产调用的旧文案路径；
+  共享服务已连接时不能仍显示整体未连接，读取故障需单独表达。
+- FIRST-12：文件身份变化、监听订阅替换、目录切换和停止后的异步结果不得
+  带入旧任务上下文；历史恢复以 session + turn 为单位。
+- FIRST-13：Hook 操作互斥且可追踪，移除使在途旧事件失效；自动检查持续只读。
+- FIRST-14：按照发现、恢复、状态/显示、宿主副作用划分代码职责；测试应调用
+  实际生产分支并覆盖反例，不能靠保留死分支维持通过率。
+
+审计发现、修复和验证统一记录于 [连接链路审计](quotaview-connection-audit-0.4.8.md)。
+本轮仍是未发布开发迭代，版本为 0.4.8 Build 4 / internal 26；不变更公开更新渠道。
+
+Build 4 最终验证：Swift 221 项、219 通过、2 项既有 opt-in 跳过、0 失败；
+连接与恢复 24 项。Universal Release 与永久开发台构建通过，App/Widget 为
+0.4.8 Build 4 / internal 26，四个主要二进制均为双架构。版本、资源、源码
+指纹及日志保存在 `dist/verification/build4/`。没有重启当前应用或替换正式安装；
+视觉、交互与真实 Intel 首装仍等待用户验收。
